Se agrega controlador de authenticacion, se mueven los endpoint correspondientes al paquete y se crea el endpoint de auth/me.
