+++ src/main/java/com/alkemy/ong/controllers/UserController.java
@@ -1,15 +1,13 @@
 package com.alkemy.ong.controllers;
 
-import com.alkemy.ong.dto.request.user.UserRegisterDto;
-import javax.validation.Valid;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.authentication.BadCredentialsException;
-import org.springframework.web.bind.annotation.*;
-import com.alkemy.ong.dto.request.user.UserLoginDto;
-import com.alkemy.ong.models.UserEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
 import com.alkemy.ong.services.UserService;
-import com.alkemy.ong.services.impl.UserServiceImpl;
 
 import lombok.RequiredArgsConstructor;
 
@@ -20,40 +18,7 @@ public class UserController {
 
     public static final String DELETE_USER = "Usuario eliminado";
     public static final String NO_DELETE_USER = "Usuario no eliminado";
-    private final UserServiceImpl userServiceImpl;
     private final UserService userService;
-
-    @PostMapping("/login")
-    public ResponseEntity<String> login(@Valid @RequestBody UserLoginDto userLoginDto) {
-        try {
-            if (userServiceImpl.findByEmail(userLoginDto.getEmail()).isPresent()) {
-                userServiceImpl.login(userLoginDto);
-                // return token
-            }else
-            	return ResponseEntity.notFound().build();
-        } catch (BadCredentialsException e) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("false");
-        }
-        return ResponseEntity.ok("ok");
-    }
-
-    /**
-     * SIGNUP Registration method that generates a token and an automatic login.
-     *
-     * @param userDTO
-     * @return String jwt Token
-     */
-    @PostMapping("/register")
-    public ResponseEntity<UserEntity> signup(@RequestBody @Valid UserRegisterDto userDTO) {
-//        userService.saveUser(userDTO);
-//        Authentication auth;
-//        auth = authenticationManager.authenticate(
-//                new UsernamePasswordAuthenticationToken(userDTO.getEmail(), userDTO.getPassword()));
-//        final String jwt = jwtTokenUtil.generateToken(auth);
-//        return ResponseEntity.ok(new AuthResponseDTO(jwt));
-        UserEntity user = userService.saveUser(userDTO);
-        return ResponseEntity.status(HttpStatus.CREATED).body(user);
-    }
 
     @DeleteMapping("user/{id}")
     public ResponseEntity<String> deleteUser(@PathVariable String id) {
